Move controls/modals tests into matching subdirectories
Aligns __tests__/components/controls/ and __tests__/components/modals/ with
src/components/controls/ and src/components/modals/. Updates relative import
paths accordingly.
