+++ src/__tests__/components/controls/ScriptureNavControls.test.tsx
@@ -6,7 +6,7 @@ import type { SerializedVerseRef } from '@sillsdev/scripture';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { useLocalizedStrings, useRecentScriptureRefs } from '@papi/frontend/react';
-import ScriptureNavControls from '../../components/controls/ScriptureNavControls';
+import ScriptureNavControls from '../../../components/controls/ScriptureNavControls';
 
 const defaultScrRef: SerializedVerseRef = { book: 'GEN', chapterNum: 1, verseNum: 1 };
